Refactor scheduler and introduced new memory handling structures
- Updated scheduler to ditch anonymous task-based approach and conform instead to a more conventional UNIXlike process/thread model
- Added a new waitpid system call to handle process waiting functionality
- Introduced a new header file for system types to define ssize_t, pid_t, and tid_t
- Implemented KVA (Kernel Virtual Address) management with functions to register, map, and retrieve KVA regions for crucial structures such as thread and process pools
- Created a pool management system for threads and processes, including initialization and stack management.
- Added error handling for KVA region mapping and ensured good-enough-for-now allocation for threads/processes/kstacks.

+++ kernel/kmem.c
@@ -58,7 +58,7 @@ static inline void bm_clear(uint32_t f) { pagebitmap[f >> 3] &= ~(1u << (f & 7))
 // Forward declaration for paging function defined in paging.c. We just want kmain to call
 // mem_init to perform all-up memory management init, but we don't expose paging_init_identity_window
 // in headers because it needs to be called exactly once.
-extern uint32_t paging_init_identity_window(uint32_t identity_bytes, paging_status_t* out_status);
+extern uint32_t paging_init_identity_window(uint64_t identity_bytes, paging_status_t* out_status);
 
 // Forward declaration for pmm functions defined in this file
 // Initializes the physical memory manager, given a set of memory regions to track
